move work queue around
